requirements.txt
langchain
langchain-community
langchain-google-genai
pandas
sentence-transformers
# chromadb
reportlab
python-dotenv 
httpx
beautifulsoup4
playwright

numpy
aiohttp
httpx
selectolax
selenium
pymongo
python-dateutil
pyarrow
langchain-huggingface
lark